diagram's relations are sorted
